+++ Samples~/UniSnake/Scripts/Scene/Play/Commands.cs
@@ -1,0 +1,6 @@
+using UnityEngine;
+
+namespace UniSnake.Scene.Play {
+  public record MoveSnakeCommand;
+  public record SetSnakeDirectionCommand(Vector2Int direction);
+}